IGNITE-17274 Improve service name lookup performance (#10123)
Adds a Map of services by service-name in addition to the Map by service-id, and use this map for lookups instead of a linear search of the map by service-name.

This significantly reduces the startup timing of services and changes the growth in timing from exponential to linear.

(cherry picked from commit 3e5cd22187814a1cf23d719fca37d709c0c15311)